Move stats calculator binding to separate module
