+++ presto-main/src/main/java/com/facebook/presto/server/ServerMainModule.java
@@ -22,34 +22,12 @@ import com.facebook.presto.client.NodeVersion;
 import com.facebook.presto.client.ServerInfo;
 import com.facebook.presto.connector.ConnectorManager;
 import com.facebook.presto.connector.system.SystemConnectorModule;
-import com.facebook.presto.cost.AggregationStatsRule;
-import com.facebook.presto.cost.AssignUniqueIdStatsRule;
-import com.facebook.presto.cost.CoefficientBasedStatsCalculator;
-import com.facebook.presto.cost.ComposableStatsCalculator;
-import com.facebook.presto.cost.ComposableStatsCalculator.Rule;
 import com.facebook.presto.cost.CostCalculator;
 import com.facebook.presto.cost.CostCalculator.EstimatedExchanges;
 import com.facebook.presto.cost.CostCalculatorUsingExchanges;
 import com.facebook.presto.cost.CostCalculatorWithEstimatedExchanges;
 import com.facebook.presto.cost.CostComparator;
-import com.facebook.presto.cost.EnforceSingleRowStatsRule;
-import com.facebook.presto.cost.ExchangeStatsRule;
-import com.facebook.presto.cost.FilterStatsCalculator;
-import com.facebook.presto.cost.FilterStatsRule;
-import com.facebook.presto.cost.JoinStatsRule;
-import com.facebook.presto.cost.LimitStatsRule;
-import com.facebook.presto.cost.OutputStatsRule;
-import com.facebook.presto.cost.ProjectStatsRule;
-import com.facebook.presto.cost.ScalarStatsCalculator;
-import com.facebook.presto.cost.SelectingStatsCalculator;
-import com.facebook.presto.cost.SelectingStatsCalculator.New;
-import com.facebook.presto.cost.SemiJoinStatsRule;
-import com.facebook.presto.cost.SimpleFilterProjectSemiJoinStatsRule;
-import com.facebook.presto.cost.StatsCalculator;
-import com.facebook.presto.cost.StatsNormalizer;
-import com.facebook.presto.cost.TableScanStatsRule;
-import com.facebook.presto.cost.UnionStatsRule;
-import com.facebook.presto.cost.ValuesStatsRule;
+import com.facebook.presto.cost.StatsCalculatorModule;
 import com.facebook.presto.event.query.QueryMonitor;
 import com.facebook.presto.event.query.QueryMonitorConfig;
 import com.facebook.presto.execution.LocationFactory;
@@ -404,8 +382,7 @@ public class ServerMainModule
         binder.bind(Metadata.class).to(MetadataManager.class).in(Scopes.SINGLETON);
 
         // statistics calculator
-        binder.bind(StatsCalculator.class).annotatedWith(SelectingStatsCalculator.Old.class).to(CoefficientBasedStatsCalculator.class).in(Scopes.SINGLETON);
-        binder.bind(StatsCalculator.class).to(SelectingStatsCalculator.class).in(Scopes.SINGLETON);
+        binder.install(new StatsCalculatorModule());
 
         // cost calculator
         binder.bind(CostCalculator.class).to(CostCalculatorUsingExchanges.class).in(Scopes.SINGLETON);
@@ -503,34 +480,6 @@ public class ServerMainModule
 
         // cleanup
         binder.bind(ExecutorCleanup.class).in(Scopes.SINGLETON);
-    }
-
-    @Provides
-    @Singleton
-    @New
-    public static StatsCalculator createNewStatsCalculator(Metadata metadata)
-    {
-        StatsNormalizer normalizer = new StatsNormalizer();
-        ScalarStatsCalculator scalarStatsCalculator = new ScalarStatsCalculator(metadata);
-        FilterStatsCalculator filterStatsCalculator = new FilterStatsCalculator(metadata, scalarStatsCalculator, normalizer);
-
-        ImmutableList.Builder<Rule<?>> rules = ImmutableList.builder();
-        rules.add(new OutputStatsRule());
-        rules.add(new TableScanStatsRule(metadata, normalizer));
-        rules.add(new SimpleFilterProjectSemiJoinStatsRule(normalizer, filterStatsCalculator)); // this must be before FilterStatsRule
-        rules.add(new FilterStatsRule(filterStatsCalculator));
-        rules.add(new ValuesStatsRule(metadata));
-        rules.add(new LimitStatsRule(normalizer));
-        rules.add(new EnforceSingleRowStatsRule(normalizer));
-        rules.add(new ProjectStatsRule(scalarStatsCalculator, normalizer));
-        rules.add(new ExchangeStatsRule(normalizer));
-        rules.add(new JoinStatsRule(filterStatsCalculator, normalizer));
-        rules.add(new AggregationStatsRule(normalizer));
-        rules.add(new UnionStatsRule(normalizer));
-        rules.add(new AssignUniqueIdStatsRule());
-        rules.add(new SemiJoinStatsRule());
-
-        return new ComposableStatsCalculator(rules.build());
     }
 
     @Provides
